# openapi/public-openapi.json
{
  "components": {
    "schemas": {
      "APIErrorResponse": {
        "properties": {
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error_code": {
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "error_code",
          "message"
        ],
        "title": "APIErrorResponse",
        "type": "object"
      },
...
      "ChangeRecord": {
        "description": "Schema for a single change record.",
        "properties": {
          "change_type": {
            "description": "Type of change: node_created, node_updated, node_deleted, edge_created, edge_updated, edge_deleted",
            "title": "Change Type",
            "type": "string"
          },
          "entity_id": {
            "description": "ID of the changed entity",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Label/type of the entity",
            "title": "Entity Label"
          },
          "entity_type": {
            "description": "node or edge",
            "title": "Entity Type",
            "type": "string"
          },
          "new_values": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "New property values",
            "title": "New Values"
          },
          "old_values": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Old property values",
            "title": "Old Values"
          },
          "properties_changed": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "List of property keys that changed",
            "title": "Properties Changed"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reason for change: user_correction, conflict_resolution, auto_update, etc.",
            "title": "Reason"
          },
          "source_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source text that triggered the change",
            "title": "Source Text"
          },
          "timestamp": {
            "description": "When the change occurred",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "change_type",
          "entity_type",
          "entity_id",
          "timestamp"
        ],
        "title": "ChangeRecord",
        "type": "object"
      },
      "ChangesResponse": {
        "description": "Response schema for changes in a time range.",
        "properties": {
          "changes": {
            "description": "List of changes",
            "items": {
              "$ref": "#/components/schemas/ChangeRecord"
            },
            "title": "Changes",
            "type": "array"
          },
          "from_timestamp": {
            "description": "Start of time range",
            "title": "From Timestamp",
            "type": "string"
          },
          "to_timestamp": {
            "description": "End of time range",
            "title": "To Timestamp",
            "type": "string"
          },
          "total_changes": {
            "description": "Total number of changes",
            "title": "Total Changes",
            "type": "integer"
          }
        },
        "required": [
          "from_timestamp",
          "to_timestamp",
          "total_changes",
          "changes"
        ],
        "title": "ChangesResponse",
        "type": "object"
      },
...
      "GQLQueryRequest": {
        "description": "Request schema for GQL/Cypher query execution.",
        "properties": {
          "format": {
            "default": "dict",
            "description": "Output format: dict, json, csv, table",
            "pattern": "^(dict|json|csv|table)$",
            "title": "Format",
            "type": "string"
          },
          "limit": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Maximum number of results (overrides query LIMIT)",
            "title": "Limit"
          },
          "parameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Query parameters (future use)",
            "title": "Parameters"
          },
          "query": {
            "description": "GQL/Cypher query string",
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "GQLQueryRequest",
        "type": "object"
      },
      "GQLQueryResponse": {
        "description": "Response schema for GQL/Cypher query execution.",
        "properties": {
          "data": {
            "description": "Query results (format depends on request)",
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error message if failed",
            "title": "Error"
          },
          "metadata": {
            "additionalProperties": true,
            "description": "Query metadata (row count, columns, etc.)",
            "title": "Metadata",
            "type": "object"
          },
          "success": {
            "description": "Whether query executed successfully",
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "GQLQueryResponse",
        "type": "object"
      },
...
      "GraphStatsResponse": {
        "description": "Response schema for comprehensive graph statistics.",
        "properties": {
          "centrality": {
            "$ref": "#/components/schemas/CentralityMetrics",
            "description": "Centrality and importance metrics"
          },
          "clustering": {
            "$ref": "#/components/schemas/ClusteringMetrics",
            "description": "Clustering and community metrics"
          },
          "current_edges": {
            "description": "Number of current edges",
            "title": "Current Edges",
            "type": "integer"
          },
          "current_nodes": {
            "description": "Number of current nodes",
            "title": "Current Nodes",
            "type": "integer"
          },
          "duplicate_pairs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DuplicateNodePair"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Actual duplicate node pairs (if requested)",
            "title": "Duplicate Pairs"
          },
          "edge_types": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Edge count by type",
            "title": "Edge Types",
            "type": "object"
          },
          "health": {
            "$ref": "#/components/schemas/HealthMetrics",
            "description": "Health and quality metrics"
          },
          "historical_edges": {
            "description": "Number of historical edges",
            "title": "Historical Edges",
            "type": "integer"
          },
          "historical_nodes": {
            "description": "Number of historical nodes",
            "title": "Historical Nodes",
            "type": "integer"
          },
          "node_types": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Node count by type",
            "title": "Node Types",
            "type": "object"
          },
          "storage_size_mb": {
            "description": "Approximate storage size in MB",
            "title": "Storage Size Mb",
            "type": "number"
          },
          "structure": {
            "$ref": "#/components/schemas/StructureMetrics",
            "description": "Graph structure metrics"
          },
          "temporal": {
            "$ref": "#/components/schemas/TemporalMetrics",
            "description": "Temporal activity metrics"
          },
          "total_edges": {
            "description": "Total number of edges",
            "title": "Total Edges",
            "type": "integer"
          },
          "total_node_properties": {
            "description": "Total number of node properties across all nodes",
            "title": "Total Node Properties",
            "type": "integer"
          },
          "total_nodes": {
            "description": "Total number of nodes",
            "title": "Total Nodes",
            "type": "integer"
          },
          "user_id": {
            "description": "User ID",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "total_nodes",
          "current_nodes",
          "historical_nodes",
          "total_edges",
          "current_edges",
          "historical_edges",
          "total_node_properties",
          "node_types",
          "edge_types",
          "storage_size_mb",
          "user_id",
          "health",
          "structure",
          "temporal",
          "centrality",
          "clustering"
        ],
        "title": "GraphStatsResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AICO Backend REST API with clean architecture",
    "title": "AICO Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/kg/changes": {
      "get": {
        "description": "Get all changes (creates, updates, deletes) in a time range.\n\n**Authentication required:** Bearer token\n\nReturns a feed of all changes to nodes and edges between two timestamps.\nUseful for building activity feeds and change logs.\n\nArgs:\n    from_timestamp: Start of time range (ISO 8601)\n    to_timestamp: End of time range (ISO 8601)\n    limit: Maximum number of changes to return\n    user: Current authenticated user (injected)\n    db_connection: Database connection (injected)\n\nReturns:\n    ChangesResponse with list of changes in time range\n    \nRaises:\n    HTTPException: 401 if unauthorized, 500 on error",
        "operationId": "get_changes_api_v1_kg_changes_get",
        "parameters": [
          {
            "description": "Start timestamp (ISO 8601)",
            "in": "query",
            "name": "from_timestamp",
            "required": true,
            "schema": {
              "description": "Start timestamp (ISO 8601)",
              "title": "From Timestamp",
              "type": "string"
            }
          },
          {
            "description": "End timestamp (ISO 8601)",
            "in": "query",
            "name": "to_timestamp",
            "required": true,
            "schema": {
              "description": "End timestamp (ISO 8601)",
              "title": "To Timestamp",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Changes",
        "tags": [
          "knowledge-graph",
          "temporal"
        ]
      }
    },
...
    "/api/v1/kg/edges": {
      "get": {
        "description": "List knowledge graph edges for the authenticated user.\n\n**Authentication required:** Bearer token\n\nQuery Parameters:\n    - limit: Number of results (default: 100, max: 1000)\n    - offset: Pagination offset (default: 0)\n\nReturns:\n    List of edges with all properties",
        "operationId": "list_edges_api_v1_kg_edges_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Edges",
        "tags": [
          "knowledge-graph"
        ]
      }
    },
    "/api/v1/kg/nodes": {
      "get": {
        "description": "List knowledge graph nodes for the authenticated user.\n\n**Authentication required:** Bearer token\n\nQuery Parameters:\n    - limit: Number of results (default: 100, max: 1000)\n    - offset: Pagination offset (default: 0)\n\nReturns:\n    List of nodes with all properties",
        "operationId": "list_nodes_api_v1_kg_nodes_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Nodes",
        "tags": [
          "knowledge-graph"
        ]
      }
    },
...
    "/api/v1/kg/query": {
      "post": {
        "description": "Execute a GQL/Cypher query against the user's knowledge graph.\n\n**Authentication required:** Bearer token\n\n**Query Examples:**\n```cypher\n# Find all people\nMATCH (p:PERSON) RETURN p.name\n\n# Find work relationships\nMATCH (p:PERSON)-[:WORKS_FOR]->(c:ORGANIZATION)\nRETURN p.name, c.name\n\n# Multi-hop traversal\nMATCH (a)-[]->(b)-[]->(c)\nWHERE a.type = 'PERSON'\nRETURN a, b, c\n```\n\n**Security:**\n- All queries automatically scoped to authenticated user\n- Query validation prevents injection attacks\n- Execution timeouts prevent DoS\n- Result size limits prevent memory exhaustion\n\nArgs:\n    request: GQL query request\n    user: Authenticated user (injected)\n    kg_storage: KG storage instance (injected)\n    db_connection: Database connection (injected)\n    \nReturns:\n    Query results with metadata\n    \nRaises:\n    HTTPException: If query execution fails",
        "operationId": "execute_gql_query_api_v1_kg_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GQLQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GQLQueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Execute Gql Query",
        "tags": [
          "knowledge-graph"
        ]
      }
    },
    "/api/v1/kg/query-templates": {
      "get": {
        "description": "Get available GQL query templates.\n\nTemplates are loaded from the OS-specific AICO data directory:\n- macOS: ~/Library/Application Support/aico/data/gql_query_templates.json\n- Linux: ~/.local/share/aico/data/gql_query_templates.json\n- Windows: %APPDATA%/aico/data/gql_query_templates.json\n\n**Authentication required:** Bearer token",
        "operationId": "get_query_templates_api_v1_kg_query_templates_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Query Templates",
        "tags": [
          "knowledge-graph"
        ]
      },
      "put": {
        "description": "Update GQL query templates (Studio template editor).\n\nAllows users to customize query templates through the Studio UI.\nTemplates are persisted to the OS-specific data directory.\n\n**Request Body:**\n```json\n{\n  \"templates\": [\n    {\n      \"id\": \"custom-query\",\n      \"title\": \"My Custom Query\",\n      \"description\": \"Description\",\n      \"category\": \"exploration\",\n      \"query\": \"MATCH (n) RETURN n LIMIT 10\",\n      \"tags\": [\"custom\"]\n    }\n  ]\n}\n```\n\n**Authentication required:** Bearer token",
        "operationId": "update_query_templates_api_v1_kg_query_templates_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Templates Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Query Templates",
        "tags": [
          "knowledge-graph",
          "Knowledge Graph"
        ]
      }
    },
    "/api/v1/kg/schema": {
      "get": {
        "description": "Get knowledge graph schema for autocomplete.\n\n**Authentication required:** Bearer token\n\nReturns:\n    Schema with node labels, relationship types, and properties",
        "operationId": "get_kg_schema_api_v1_kg_schema_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Kg Schema Api V1 Kg Schema Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Kg Schema",
        "tags": [
          "knowledge-graph"
        ]
      }
    },
    "/api/v1/kg/stats": {
      "get": {
        "description": "Get statistics about the user's knowledge graph.\n\n**Authentication required:** Bearer token\n\nReturns:\n    Graph statistics including node/edge counts and type distributions\n    \nRaises:\n    HTTPException: If stats retrieval fails",
        "operationId": "get_graph_stats_api_v1_kg_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Graph Stats",
        "tags": [
          "knowledge-graph"
        ]
      }
    },
...
  }
}